susClient-master_cardiff/susClient-master/AUS/requirements.txt
pywinauto==0.6.8
selenium==3.141.0
psutil==5.7.0
requests==2.23.0
msedge-selenium-tools==3.141.2
PyAutoGUI==0.9.50